tkinterKood/AP/requirements.txt
numpy == 1.23.4
matplotlib == 3.6.2
tensorflow == 2.12.0rc1
keras == 2.12.0rc1
Pillow == 9.3.0
qbstyles == 0.1.4







